refactor: post.scheme.ts 파일명을 post.schema.ts로 수정하고 스키마 변수명 정리
