+++ src/features/auth/ui/LoginForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SocialLoginButtons } from './SocialLoginButtons';
 import { InputWithLabel } from '@/shared/ui/form/InputWithLabel';
-import { loginSchema, type LoginFormValues } from '../model/auth.schema';
+import { loginSchema, type LoginFormValues } from '../lib/auth.schema';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useLogin } from '../api/useLogin';
