+++ src/features/auth/ui/SignupForm.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react';
 import { SocialLoginButtons } from './SocialLoginButtons';
 import { InputWithLabel } from '@/shared/ui/form/InputWithLabel';
 import { useForm } from 'react-hook-form';
-import { signupSchema, type SignupFormValues } from '../model/auth.schema';
+import { signupSchema, type SignupFormValues } from '../lib/auth.schema';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useSignup } from '../api/useSignup';
 import { Check, Clock } from 'lucide-react';
